# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: isachore API
  description: |
    Chore management for households.

    ## Authenticating

    Two transports, either of which carries a session:

    - the httpOnly `isachore_token` cookie, which a browser sends on its own, and
    - the same opaque token as an `Authorization: Bearer` header, for clients with no cookie jar.

    Two ways to open one, and nothing downstream can tell them apart. `POST /api/v1/auth/login`
    takes an email and password, and answers `two_factor_required` instead of a session when the
    account has TOTP enabled - finish that with `POST /api/v1/auth/verify-2fa`. Where single
    sign-on is configured, `GET /api/v1/auth/oidc/start` is the other way in, and on a server
    with `OIDC_ONLY` set it is the only one: password login answers 403 there.
    `GET /api/v1/auth/methods` says which of the two this server offers, without a session.

    There is no self-registration: administrators create accounts, and the first administrator
    comes from the `init` CLI command.

    ## The CSRF header

    A request authenticated **by cookie** that uses an unsafe method (POST, PATCH, PUT, DELETE)
    must also carry a non-empty `X-CSRF-Token` header, or it is refused with 403 before it
    reaches the route. Any value will do: this is a custom-header defence over `SameSite=Lax`,
    so what matters is that a cross-site form cannot set the header at all.

...
  version: 0.1.0
paths:
...
  /api/v1/households:
    get:
      tags:
        - households
      summary: List Households
      operationId: list_households_api_v1_households_get
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Page Size
        - name: sort_by
          in: query
          required: false
          schema:
            enum:
              - id
              - name
              - created_at
            type: string
            default: created_at
            title: Sort By
        - name: sort_dir
          in: query
          required: false
          schema:
            enum:
              - asc
              - desc
            type: string
            default: desc
            title: Sort Dir
        - name: name
          in: query
          required: false
          schema:
            anyOf:
              - type: string
                maxLength: 255
              - type: 'null'
            title: Name
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_HouseholdListRead_'
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
        - households
      summary: Create Household
      operationId: create_household_api_v1_households_post
      security:
        - sessionCookie: []
        - bearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HouseholdCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdListRead'
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/households/{household_id}:
    get:
      tags:
        - households
      summary: Get Household
      operationId: get_household_api_v1_households__household_id__get
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - name: household_id
          in: path
          required: true
          schema:
            type: integer
            title: Household Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdListRead'
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: No household with this id that you are a member of.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
        - households
      summary: Update Household
      operationId: update_household_api_v1_households__household_id__patch
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - name: household_id
          in: path
          required: true
          schema:
            type: integer
            title: Household Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HouseholdUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdListRead'
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Only the household's owner may do this. Being an organiser is not enough, and no role change grants it - ownership has to be transferred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: No household with this id that you are a member of.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '409':
          description: A chore was completed while the timezone was being changed, so a rescheduled occurrence collided with the closed one. Retrying is safe.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - households
      summary: Delete Household
      operationId: delete_household_api_v1_households__household_id__delete
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - name: household_id
          in: path
          required: true
          schema:
            type: integer
            title: Household Id
      responses:
        '204':
          description: Successful Response
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Only the household's owner may do this. Being an organiser is not enough, and no role change grants it - ownership has to be transferred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: No household with this id that you are a member of.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/households/{household_id}/members:
    get:
      tags:
        - households
      summary: List Household Members
      operationId: list_household_members_api_v1_households__household_id__members_get
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - name: household_id
          in: path
          required: true
          schema:
            type: integer
            title: Household Id
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Page Size
        - name: sort_by
          in: query
          required: false
          schema:
            enum:
              - id
              - name
            type: string
            default: name
            title: Sort By
        - name: sort_dir
          in: query
          required: false
          schema:
            enum:
              - asc
              - desc
            type: string
            default: asc
            title: Sort Dir
        - name: name
          in: query
          required: false
          schema:
            anyOf:
              - type: string
                maxLength: 255
              - type: 'null'
            title: Name
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_HouseholdMemberRoleRead_'
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: No household with this id that you are a member of.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/households/{household_id}/members/{user_id}:
    patch:
      tags:
        - households
      summary: Update Household Member
      description: |-
        Set one member's role. The owner may set any of the three; an organiser may only move
        people between deputy and helper.

        That asymmetry is the point: an organiser can share the day-to-day load without being able
        to grow the set of people who could demote them. So they may not hand out `organiser`, may
        not touch a row that already holds it, and therefore may not demote themselves either -
        that last one falls out of the same rule rather than needing its own check.

        The checks are ordered deliberately. The owner's row is refused with a 409 **before** any
        caller-specific rule, because "the owner is always an organiser" is a property of the
        target: everyone gets the same actionable answer (transfer ownership, which promotes the
        new owner) instead of an organiser getting a 403 about a rule that is not the real reason.
      operationId: update_household_member_api_v1_households__household_id__members__user_id__patch
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - name: household_id
          in: path
          required: true
          schema:
            type: integer
            title: Household Id
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HouseholdMemberUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdMemberRoleRead'
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Either the role held in this household does not reach organiser, or the change involves the organiser role itself, which only the household's owner may grant or alter. The second cannot be cleared by a promotion, only by a transfer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: No household with this id that you are a member of, or no active member of it with this user id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '409':
          description: That member owns the household, and an owner is always an organiser. Transfer ownership instead.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - households
      summary: Remove Household Member
      operationId: remove_household_member_api_v1_households__household_id__members__user_id__delete
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - name: household_id
          in: path
          required: true
          schema:
            type: integer
            title: Household Id
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            title: User Id
      responses:
        '204':
          description: Successful Response
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Only the household's owner may do this. Being an organiser is not enough, and no role change grants it - ownership has to be transferred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: No household with this id that you are a member of, or nobody with this user id belongs to it. Unlike setting a role, this reaches a member whose account is disabled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '409':
          description: That member owns the household. Transfer ownership before removing them.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/admin/households:
    get:
      tags:
        - admin-households
      summary: List Households
      operationId: list_households_api_v1_admin_households_get
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Page Size
        - name: sort_by
          in: query
          required: false
          schema:
            enum:
              - id
              - name
              - created_at
            type: string
            default: created_at
            title: Sort By
        - name: sort_dir
          in: query
          required: false
          schema:
            enum:
              - asc
              - desc
            type: string
            default: desc
            title: Sort Dir
        - name: name
          in: query
          required: false
          schema:
            anyOf:
              - type: string
                maxLength: 255
              - type: 'null'
            title: Name
        - name: status
          in: query
          required: false
          schema:
            enum:
              - active
              - deleted
              - all
            type: string
            default: active
            title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_HouseholdListRead_'
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Signed in, but not a site administrator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
        - admin-households
      summary: Create Household
      operationId: create_household_api_v1_admin_households_post
      security:
        - sessionCookie: []
        - bearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HouseholdCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdListRead'
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Signed in, but not a site administrator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/households/{household_id}:
    get:
      tags:
        - admin-households
      summary: Get Household
      operationId: get_household_api_v1_admin_households__household_id__get
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - name: household_id
          in: path
          required: true
          schema:
            type: integer
            title: Household Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdListRead'
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Signed in, but not a site administrator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: No household with this id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
        - admin-households
      summary: Update Household
      operationId: update_household_api_v1_admin_households__household_id__patch
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - name: household_id
          in: path
          required: true
          schema:
            type: integer
            title: Household Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HouseholdUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdListRead'
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Signed in, but not a site administrator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: No household with this id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '409':
          description: A chore was completed while this household's timezone was being changed, so a rescheduled occurrence collided with the closed one. Retrying is safe.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - admin-households
      summary: Delete Household
      operationId: delete_household_api_v1_admin_households__household_id__delete
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - name: household_id
          in: path
          required: true
          schema:
            type: integer
            title: Household Id
      responses:
        '204':
          description: Successful Response
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Signed in, but not a site administrator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: No household with this id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Page_HouseholdListRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/HouseholdListRead'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
      type: object
      required:
        - items
        - total
        - page
        - page_size
      title: Page[HouseholdListRead]
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      description: The httpOnly session cookie, set by `POST /api/v1/auth/login` or by the single sign-on callback. A browser sends it automatically; an unsafe method additionally needs the `X-CSRF-Token` header.
      in: cookie
      name: isachore_token
    bearerToken:
      type: http
      description: 'The same opaque session token as an `Authorization: Bearer` header, for clients with no cookie jar. Exempt from the `X-CSRF-Token` requirement.'
      scheme: bearer
    parkedAdminCookie:
      type: apiKey
      description: An impersonating administrator's own session, parked while they act as somebody else. `POST /api/v1/auth/stop-impersonating` is the only route it *authenticates* - several others read it to attribute an action to the real operator, and logout reads it to end both sessions.
      in: cookie
      name: isachore_admin_token